read a script file with calls to chown via -F [filename] or -
basically this patch is dedicated to read a script
file with loads of calls to chown eg. like:

chown 0:0 /path/to/file

If chown would be called from within the script that would spawn
a new process everytime and consumes a lot of time. Hence we read
read such a script file as regular file or from stdin and feed it
internally.